+++ docs/changelog.rst
@@ -5,6 +5,11 @@ Changelog
 
     Version 4 introduces breaking changes.  Please refer to :doc:`release notes<release_notes>`.
 
+4.0.3 ()
+------------------
+
+- Added logging details to configuration (`1821 <https://github.com/django-import-export/django-import-export/pull/1821>`_)
+
 4.0.2 (2024-05-13)
 ------------------